Create blink_detect.py
I have implemented a new version to detect blink. I have performed a Canny Edge Detection on the image and later have tried to make contours and then through checking the length of the hull or contours, we detect the blink.

if the length < 3 then it is a blink else no blink. the method has been tested on two videos and results are very promising better than the previous approach.